Was added a opprortunity to attach photo to message and other minor changes/

+++ WeatherBot/ClassesforWeatherInfo/Main.cs
@@ -8,7 +8,7 @@ namespace WeatherBot.ClassesforWeatherInfo
     {
         public double temp { get; set; }
         public double feels_like { get; set; }
-        public int temp_min { get; set; }
+        public double temp_min { get; set; }
         public double temp_max { get; set; }
         public int pressure { get; set; }
         public int humidity { get; set; }
